feat: add User.js
